+++ src/domain/protocols/ai/aiClient.protocol.ts
