+++ src/domain/protocols/ai/index.ts
