+++ src/domain/services/ai.service.ts
@@ -1,5 +1,3 @@
-import { ListMoviesEntity, MovieEntity } from "../entities";
-
 export interface AiService {
   generateResponse: (prompt: AiService.Params) => Promise<AiService.Model>;
 }
