+++ src/infrastructure/helpers/movieEntity.helper.ts
