+++ src/main/factories/usecases/getMovie.factory.ts
@@ -1,6 +1,6 @@
 import { GetMovieUsecase } from "@/domain/usecases";
-import { GetMovieUsecaseImpl } from "@/usecases";
 import { makeMoviesRepository } from "../repositories/movies.factory";
+import { GetMovieUsecaseImpl } from "@/infrastructure/usecases";
 
 export const makeGetMovieUsecase = (): GetMovieUsecase =>
   new GetMovieUsecaseImpl(makeMoviesRepository());
