+++ src/main/factories/usecases/getMovieSuggestion.factory.ts
@@ -1,8 +1,8 @@
 import { GetMovieSuggestionUsecase } from "@/domain/usecases";
 
 import { makeMoviesRepository } from "../repositories/movies.factory";
-import { GetMovieSuggestionUsecaseImpl } from "@/usecases/getMovieSuggestion.usecase";
 import { makeAiService } from "../services/ai.factory";
+import { GetMovieSuggestionUsecaseImpl } from "@/infrastructure/usecases/getMovieSuggestion.usecase";
 
 export const makeGetMovieSuggestionUsecase = (): GetMovieSuggestionUsecase =>
   new GetMovieSuggestionUsecaseImpl(makeMoviesRepository(), makeAiService());
